dom/media/mediasource/test/test_LoadedMetadataFired.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: append initialization only</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/webm");

  sb.appendBuffer(new Uint8Array(await fetchWithXHR("seek.webm"), 0, 318));
  v.play();
  await once(v, "loadedmetadata");
  ok(true, "Got loadedmetadata event");
  is(v.videoWidth, 320, "videoWidth has correct initial value");
  is(v.videoHeight, 240, "videoHeight has correct initial value");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
